Separate segment window flush manager
